--- frontend/components/session_form/signup_form_container.jsx
@@ -1,21 +1,0 @@
-import { connect } from 'react-redux';
-import React from 'react';
-import { Link } from 'react-router-dom';
-import { signup } from '../../actions/session_actions';
-import SignupForm from './signup_form';
-
-const mapStateToProps = ({ errors }) => {
-    return {
-        errors: errors.session,
-        formType: 'Sign Up',
-        navText: "Already have an account?"
-    };
-};
-
-const mapDispatchToProps = dispatch => {
-    return {
-        processForm: (user) => dispatch(signup(user)),
-    };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(SignupForm);
